feat: enrich query results with manifest cards
